Create grafana_update.py
This script updates the grafana container. I wanted to get around to obtaining a list of all the image IDs and then go through the array to determine which local images are downloaded and which are to be removed but I never got around to this as my docker containers update automatically already.